requirements.txt
amqp==5.3.1
anyio==4.12.1
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bcrypt==5.0.0
billiard==4.2.4
brotli==1.2.0
CacheControl==0.14.4
cbor2==5.7.1
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
Django==5.2.9
django-celery==3.1.17
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
exceptiongroup==1.3.1
filelock==3.20.3
firebase_admin==7.1.0
fonttools==4.61.1
freetype-py==2.5.1
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
lxml==6.0.2
msgpack==1.1.2
oscrypto==1.3.0
packaging==25.0
pdfkit==1.0.0
pillow==12.0.0
playwright==1.57.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydyf==0.12.1
pyee==13.0.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.5.0
pyphen==0.17.2
python-bidi==0.6.7
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.30.0
rsa==4.9.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==8.1.1
zopfli==0.4.0
